rename "api" directories to "types"
